segmentation_pspnet/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2021.5.30
charset-normalizer==2.0.6; python_version >= '3'
cycler==0.10.0
idna==3.2; python_version >= '3'
kiwisolver==1.3.2; python_version >= '3.7'
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
pillow==8.3.2
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
requests==2.26.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
torch==1.9.1
typing-extensions==3.10.0.2
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
